pyproject.toml
[tool.poetry]
name = "saleor"
version = "3.12.7"
description = "A modular, high performance, headless e-commerce platform built with Python, GraphQL, Django, and React."
authors = [ "Saleor Commerce <hello@saleor.io>" ]
license = "BSD-3-Clause"
readme = "README.md"
homepage = "https://saleor.io/"
repository = "https://github.com/mirumee/saleor"
documentation = "https://docs.saleor.io/"

  [tool.poetry.dependencies]
  python = "~3.9"
  babel = ">=2.8,<2.12"
  boto3 = "^1.26"
  braintree = ">=4.2,<4.19"
  dj-database-url = "^1"
  dj-email-url = "^1"
  django-countries = "^7.2"
  django-filter = "^22.1"
  django-measurement = "^3.0"
  django-mptt = "^0"
  django-phonenumber-field = ">=4,<8"
  django-prices = "^2.3"
  django-prices-openexchangerates = "^1.2.0"
  django-prices-vatlayer = "^1.1.0"
  django-stubs-ext = "^0.7.0"
  django-versatileimagefield = "^2.1"
  draftjs-sanitizer = "^1.0.0"
  faker = ">=4.1,<19.0"
  freezegun = "^1"
  google-cloud-storage = "^2.0.0"
  google-i18n-address = "^2.5.2"
  google-measurement-protocol = "^1.0"
  html-to-draftjs = "^1.0.1"
  markdown = "^3.1.1"
  petl = "1.7.12"
  opentracing = "^2.3.0"
  phonenumberslite = "^8.12.25"
  prices = "^1.0"
  psycopg2-binary = "^2.8.3"
  razorpay = "^1.2"
  requests = "^2.22"
  sentry-sdk = "^1.9"
  stripe = "^3.0.0"
  text-unidecode = "^1.2"
  weasyprint = ">=53.0"
  oauthlib = "^3.1"
  jaeger-client = "^4.5.0"
  openpyxl = "^3.0.3"
  django-cache-url = "^3.1.2"
  pyjwt = "2.6.0"
  python-json-logger = ">=0.1.11,<2.1.0"
  pytimeparse = "^1.1.8"
  django-redis = "^5.0.0"
  Adyen = "^4.0.0"
  google-cloud-pubsub = ">=1.7,<3.0"
  gunicorn = "^20.0.4"
  pybars3 = "^0.9.7"
  html2text = "^2020.1.16"
  authorizenet = "^1.1.3"
  azure-common = "^1.1.28"
  azure-storage-blob = "^12.12.0"
  azure-storage-common = "^2.1.0"
  sendgrid = "^6.7.1"
  micawber = "^0.5.2"
  django-celery-beat = "^2.2.1"
  posuto = "^2023.2.0"
  cryptography = "^40.0.1"
  graphene = "<3.0"
  Authlib = "^1.0.0"
  pillow-avif-plugin = "^1.3.1"
  google-cloud-secret-manager = "^2.16.1"
  django-allow-cidr = "0.6.0"
  djangorestframework = "3.14.0"

  wescommon = "0.1.0.13"
  wes-fg = "^0.1.0.6"


    [tool.poetry.dependencies.django]
    version = "^3.2.19"
    extras = [ "bcrypt" ]

    [tool.poetry.dependencies.uvicorn]
    extras = [ "standard" ]
    version = "^0.19.0"

    [tool.poetry.dependencies.celery]
    version = ">=4.4.5,<6.0.0"
    extras = [ "redis" ]

    [tool.poetry.dependencies.django-storages]
    version = "^1.11"
    extras = [ "google" ]

    [tool.poetry.dependencies.python-magic-bin]
    version = "^0.4.14"
    platform = "win32"

  [tool.poetry.dev-dependencies]
  black = "23.3.0"
  codecov = "^2.1.13"
  coverage = "^6.3"
  django-debug-toolbar = "^3.2"
  django-debug-toolbar-request-history = "^0"
  django-graphiql-debug-toolbar = "^0.2.0"
  django-extensions = "^3.1.2"
  flake8 = "^4.0.1"
  isort = "^5.12.0"
  pre-commit = "^3.0"
  pycodestyle = "^2.5"
  pydocstyle = "^6.1"
  pylint = "^2.8.3"
  pylint-celery = "^0"
  pylint-django = "^2.4.4"
  pylint-plugin-utils = "^0"
  pytest = "^7.0.1"
  pytest-asyncio = "^0.20.3"
  pytest-celery = "^0.0.0"
  pytest-cov = "^4.0.0"
  pytest-django = "4.5.2"
  pytest-django-queries = "~1.2"
  pytest-mock = "^3.6.1"
  pytest-recording = "^0.12.0"
  pytest-socket = "^0.5.1"
  pytest-xdist = "^3.0.2"
  mypy = "^1.0.0"
  mypy-extensions = "^1.0.0"
  pywatchman = "^1.4.1"
  types-mock = "^5.0.0.5"
  types-redis = "^4.2.6"
  types-requests = "^2.28.2"
  types-pytz = "^2022.1.1"
  types-pkg-resources = "^0.1.3"
  types-python-dateutil = "^2.8.19"
  django-stubs = "^1.14.0"
  before_after = "^1.0.1"
  types-certifi = "^2021.10.8"
  types-freezegun = "^1.1.7"
  types-six = "^1.16.17"
  fakeredis = "^1.8"

[tool.black]
target_version = [ "py35", "py36", "py37", "py38" ]
include = "\\.pyi?$"
exclude = """
/(\\.git/
  |\\.eggs
  |\\.hg
  |__pycache__
  |\\.cache
  |\\.ipynb_checkpoints
  |\\.mypy_cache
  |\\.pytest_cache
  |\\.tox
  |\\.venv
  |node_modules
  |_build
  |buck-out
  |build
  |dist
  |media
  |infrastructure
  |templates
  |locale
)/
"""

[[tool.poetry.source]]
name = "wellstand-pypi"
url = "https://us-python.pkg.dev/non-production-381217/wellstand-pypi/simple/"
